backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Basic utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development tools
pytest==7.4.3
httpx==0.25.2

# AI Integration
ollama==0.1.7
httpx==0.25.2
